+++ test/units/basic_promise.js
