+++ test/units/examples.js
